requirements.txt
astroid==2.4.2
Authlib==0.14.3
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
Deprecated==1.2.10
Flask==1.1.2
gunicorn==20.0.4
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
loginpass==0.5
MarkupSafe==1.1.1
mccabe==0.6.1
pycparser==2.20
PyGithub==1.51
PyJWT==1.7.1
pylint==2.5.3
requests==2.23.0
six==1.15.0
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
